+++ src/Boleto.Net/Banco/Banco_Brasil.cs
@@ -184,7 +184,7 @@ namespace BoletoNet
             #region Carteira 18
             //Carteira 18 com nosso número de 11 posições
             if (boleto.Carteira.Equals("18"))
-                boleto.NossoNumero = Utils.FormatCode(boleto.NossoNumero, 11);
+                boleto.NossoNumero = Utils.FormatCode(boleto.NossoNumero, 10);
             #endregion Carteira 18
 
             #region Carteira 18-019
@@ -644,12 +644,13 @@ namespace BoletoNet
             #region Carteira 18
             if (boleto.Carteira.Equals("18"))
             {
-                boleto.CodigoBarra.Codigo = string.Format("{0}{1}{2}{3}{4}{5}{6}",
+                boleto.CodigoBarra.Codigo = string.Format("{0}{1}{2}{3}{4}{5}{6}{7}",
                     Utils.FormatCode(Codigo.ToString(), 3),
                     boleto.Moeda,
                     FatorVencimento(boleto),
                     valorBoleto,
-                    boleto.Cedente.Codigo.PadLeft(12, '0'),
+                    "000000",
+                    boleto.Cedente.Codigo,
                     boleto.NossoNumero,
                     Utils.FormatCode(LimparCarteira(boleto.Carteira), 2));
             }
@@ -974,7 +975,7 @@ namespace BoletoNet
             D = Posição 25 a 34 do código de barras
             Y = DV que amarra o campo 2 (Módulo 10, contido no Anexo 7)
              */
-            campo2 = Strings.Mid(cmplivre, 5, 10);
+            campo2 = Strings.Mid(cmplivre, 6, 10);
             digitoMod = Mod10(campo2);
             campo2 = campo2 + digitoMod.ToString();
             campo2 = Strings.Mid(campo2, 1, 5) + "." + Strings.Mid(campo2, 6, 6);
